refactor: move device_model and device_store into pages
