+++ apps/firmware_updater/lib/pages.dart
@@ -1,3 +1,5 @@
 export 'pages/detail_page.dart';
+export 'pages/device_model.dart';
 export 'pages/device_page.dart';
+export 'pages/device_store.dart';
 export 'pages/release_page.dart';
