+++ apps/firmware_updater/lib/pages/detail_page.dart
@@ -1,5 +1,3 @@
-import 'package:firmware_updater/device_model.dart';
-import 'package:firmware_updater/device_store.dart';
 import 'package:firmware_updater/pages.dart';
 import 'package:firmware_updater/services.dart';
 import 'package:flutter/material.dart';
