+++ apps/firmware_updater/lib/pages/device_model.dart
