+++ apps/firmware_updater/lib/pages/device_store.dart
