+++ apps/firmware_updater/lib/pages/release_page.dart
@@ -1,7 +1,6 @@
-import 'package:firmware_updater/device_model.dart';
-import 'package:firmware_updater/device_store.dart';
 import 'package:firmware_updater/fwupd_notifier.dart';
 import 'package:firmware_updater/fwupd_x.dart';
+import 'package:firmware_updater/pages.dart';
 import 'package:firmware_updater/widgets/release_card.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
