+++ apps/firmware_updater/test/release_page_test.dart
@@ -1,5 +1,3 @@
-import 'package:firmware_updater/device_model.dart';
-import 'package:firmware_updater/device_store.dart';
 import 'package:firmware_updater/fwupd_notifier.dart';
 import 'package:firmware_updater/pages.dart';
 import 'package:flutter/material.dart';
