+++ apps/firmware_updater/test/test_utils.dart
@@ -1,5 +1,4 @@
-import 'package:firmware_updater/device_model.dart';
-import 'package:firmware_updater/device_store.dart';
+import 'package:firmware_updater/pages.dart';
 import 'package:firmware_updater/fwupd_notifier.dart';
 import 'package:firmware_updater/services.dart';
 import 'package:flutter/material.dart';
