+++ apps/firmware_updater/test/test_utils.mocks.dart
@@ -7,9 +7,9 @@ import 'dart:async' as _i4;
 import 'dart:io' as _i9;
 import 'dart:ui' as _i5;
 
-import 'package:firmware_updater/device_model.dart' as _i3;
-import 'package:firmware_updater/device_store.dart' as _i10;
 import 'package:firmware_updater/fwupd_notifier.dart' as _i6;
+import 'package:firmware_updater/pages/device_model.dart' as _i3;
+import 'package:firmware_updater/pages/device_store.dart' as _i10;
 import 'package:firmware_updater/services.dart' as _i8;
 import 'package:fwupd/fwupd.dart' as _i2;
 import 'package:mockito/mockito.dart' as _i1;
